fix: typos & move hooks
