+++ src/utils/index.ts
@@ -1,3 +1,3 @@
+export * from "../hooks/useClickOutside";
 export * from "./css";
 export * from "./string";
-export * from "./useClickOutside";
